+++ core/src/main/java/roomescape/auth/support/JwtProvider.java
@@ -1,9 +1,8 @@
-package roomescape.auth.application;
+package roomescape.auth.support;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import roomescape.auth.application.dto.JwtPayload;
 import roomescape.domain.common.ClockHolder;
@@ -14,24 +13,19 @@ import java.util.Date;
 @RequiredArgsConstructor
 public class JwtProvider {
 
-    @Value("${security.jwt.token.secret-key}")
-    private String secretKey;
-
-    @Value("${security.jwt.token.expiration-time}")
-    private long validityInMilliseconds;
-
     private final ClockHolder clockHolder;
+    private final JwtConfig jwtConfig;
 
     public String createToken(final JwtPayload jwtPayload) {
         final Date issuedAt = Date.from(clockHolder.getCurrentSeoulZonedDateTime().toInstant());
-        final Date expirationDate = new Date(issuedAt.getTime() + this.validityInMilliseconds);
+        final Date expirationDate = new Date(issuedAt.getTime() + jwtConfig.getValidityInMilliseconds());
 
         return Jwts.builder()
                 .claim("role", jwtPayload.role())
                 .claim("email", jwtPayload.email())
                 .issuedAt(issuedAt)
                 .expiration(expirationDate)
-                .signWith(Keys.hmacShaKeyFor(this.secretKey.getBytes()))
+                .signWith(Keys.hmacShaKeyFor(jwtConfig.getSecretKey().getBytes()))
                 .compact();
     }
 }
